# docs/openapi/spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API de Usuários",
    "version": "1.0.0",
    "description": "API REST de gestão de usuários e famílias, orientada a eventos."
  },
  "servers": [
    {
      "url": "http://localhost:8082",
      "description": "Local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/v1/profile": {
      "get": {
        "tags": ["Profile"],
        "summary": "Obtém o perfil do usuário autenticado",
        "description": "Retorna os dados do perfil. Se ainda não existir, provisiona automaticamente a partir das claims do JWT.",
        "operationId": "getProfile",
        "responses": {
          "200": {
            "description": "Perfil retornado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileResponse" },
                "example": {
                  "cpf": "123.456.789-00",
                  "email": "joao@email.com",
                  "nomeCompleto": "João Silva",
                  "telefone": "11999999999"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "put": {
        "tags": ["Profile"],
        "summary": "Atualiza o perfil do usuário autenticado",
        "description": "Atualiza nomeCompleto, cpf, email e telefone. Sincroniza nome e email com o Keycloak via Admin API. Publica o evento PROFILE_UPDATED.",
        "operationId": "updateProfile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProfileUpdateRequest" },
              "example": {
                "nomeCompleto": "João Silva",
                "cpf": "123.456.789-00",
                "email": "joao@email.com",
                "telefone": "11999999999"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "502": { "$ref": "#/components/responses/BadGateway" }
        }
      }
    },
    "/api/v1/profile/password": {
      "put": {
        "tags": ["Profile"],
        "summary": "Altera a senha do usuário autenticado",
        "description": "Valida a senha atual via ROPC e redefine a nova senha via Admin API do Keycloak. Publica o evento PROFILE_PASSWORD_CHANGED.",
        "operationId": "changePassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PasswordChangeRequest" },
              "example": {
                "senhaAtual": "senhaAntiga123",
                "novaSenha": "novaSenha456",
                "confirmarNovaSenha": "novaSenha456"
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Senha alterada com sucesso" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "502": { "$ref": "#/components/responses/BadGateway" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido via Keycloak (dev-client)."
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Dados inválidos",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "status": 400, "error": "Bad Request", "message": "As senhas não conferem" }
          }
        }
      },
      "Unauthorized": {
        "description": "Token ausente ou inválido",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "status": 401, "error": "Unauthorized", "message": "Token ausente ou inválido" }
          }
        }
      },
      "Forbidden": {
        "description": "Acesso negado",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "status": 403, "error": "Forbidden", "message": "Acesso negado" }
          }
        }
      },
      "NotFound": {
        "description": "Recurso não encontrado",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "status": 404, "error": "Not Found", "message": "Recurso não encontrado" }
          }
        }
      },
...
      "BadGateway": {
        "description": "Falha ao comunicar com o Keycloak",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "status": 502, "error": "Bad Gateway", "message": "Falha ao sincronizar dados com Keycloak" }
          }
        }
      }
    },
    "schemas": {
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "cpf":          { "type": "string", "description": "CPF do usuário no formato 000.000.000-00", "example": "123.456.789-00" },
          "email":        { "type": "string", "format": "email", "description": "Endereço de e-mail do usuário", "example": "joao@email.com" },
          "nomeCompleto": { "type": "string", "description": "Nome completo do usuário", "example": "João Silva" },
          "telefone":     { "type": "string", "description": "Telefone do usuário com DDD", "example": "11999999999" }
        }
      },
      "ProfileUpdateRequest": {
        "type": "object",
        "properties": {
          "nomeCompleto": { "type": "string", "description": "Nome completo do usuário. Sincronizado com o Keycloak.", "example": "João Silva" },
          "cpf":          { "type": "string", "description": "CPF do usuário no formato 000.000.000-00", "example": "123.456.789-00" },
          "email":        { "type": "string", "format": "email", "description": "E-mail do usuário. Sincronizado com o Keycloak.", "example": "joao@email.com" },
          "telefone":     { "type": "string", "description": "Telefone do usuário com DDD", "example": "11999999999" }
        }
      },
      "PasswordChangeRequest": {
        "type": "object",
        "required": ["senhaAtual", "novaSenha", "confirmarNovaSenha"],
        "properties": {
          "senhaAtual":         { "type": "string", "description": "Senha atual do usuário, validada via ROPC no Keycloak", "example": "senhaAntiga123" },
          "novaSenha":          { "type": "string", "description": "Nova senha desejada", "example": "novaSenha456" },
          "confirmarNovaSenha": { "type": "string", "description": "Confirmação da nova senha, deve ser idêntica a novaSenha", "example": "novaSenha456" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time", "description": "Data e hora do erro em UTC" },
          "status":    { "type": "integer", "description": "Código HTTP do erro", "example": 400 },
          "error":     { "type": "string", "description": "Descrição resumida do erro HTTP", "example": "Bad Request" },
          "message":   { "type": "string", "description": "Mensagem detalhada do erro", "example": "Mensagem de erro" },
          "path":      { "type": "string", "description": "Caminho do endpoint que gerou o erro", "example": "/api/v1/profile" }
        }
      }
    }
  }
}